% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roc.R
\name{get_auc}
\alias{get_auc}
\title{Computes the area under the curve from \code{get_roc()} output}
\usage{
get_auc(roc.)
}
\arguments{
\item{roc.}{Nx2 matrix of false and true positive rates, often the output
from \code{get_roc()}}
}
\value{
A value between zero and one which represents the area under the ROC
curve
}
\description{
Computes the area under the curve from \code{get_roc()} output
}
\examples{
predicted_probs <- nnadic(test_data)
roc <- get_roc(predicted_probabilities = predicted_prob,
               truth = test_response,
               make_plot = TRUE,
               add_auc = TRUE)
get_auc(roc)
}
