% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roc.R
\name{get_brier}
\alias{get_brier}
\title{Computes the Brier Score}
\usage{
get_brier(predicted_probabilities, truth)
}
\arguments{
\item{predicted_probabilities}{the "probabilities" returned from \code{nnadic()}}

\item{truth}{the true classifications (a vector of zeros and ones)}
}
\value{
a value between zero and one that is equivalent to the MSE
}
\description{
Computes the Brier Score
}
\examples{
predicted_probs <- nnadic(test_data)
get_brier(predicted_probabilities = predicted_prob,
               truth = test_response)
}
