% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roc.R
\name{get_confusion_matrix}
\alias{get_confusion_matrix}
\title{Computes the confusion matrix for a given threshold}
\usage{
get_confusion_matrix(pred_prob, truth., t_hold = 0.5)
}
\arguments{
\item{pred_prob}{the "probabilities" returned from \code{nnadic()}}

\item{truth.}{the true classifications (a vector of zeros and ones)}

\item{t_hold}{what the cutoff is for classifying a dataset as AI}
}
\value{
the confusion matrix as a 2x2 \code{matrix} object
}
\description{
Computes the confusion matrix for a given threshold
}
\details{
Confusion matrix is:

True Negative | False Negative

False Positive | True Positive
}
\examples{
predicted_probs <- nnadic(test_data_four)
get_confusion_matrix(pred_prob = predicted_prob, truth = test_response_four)
}
