% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_marginal.R
\name{get_nnadic_input}
\alias{get_nnadic_input}
\title{Get dataset into format needed for \code{nnadic()}'s prediction}
\usage{
get_nnadic_input(
  data,
  make_exponential = TRUE,
  subsample = TRUE,
  num_datasets = 100,
  include_all = TRUE,
  comp_lag = 1
)
}
\arguments{
\item{data}{a matrix with two columns}

\item{make_exponential}{Change to \code{FALSE} if the data
already have exponential margins. (defaults to \code{TRUE})}

\item{subsample}{Change to \code{FALSE} if the dataset has at least than 10,000
points and you want to take the top 500 points rather than subsample all
points above the 0.95 quantile. (defaults to \code{TRUE})}

\item{num_datasets}{The number of output datasets (defaults to \code{100})}

\item{include_all}{If there are fewer than 10,000 points in your data should
all of the top five percent be included in every dataset as many times as
possible (\code{TRUE}) or should all 500 points be from resampling (\code{FALSE})?
(defaults to \code{TRUE})}

\item{comp_lag}{If your data is a univariate time series the function will
automatically lag it so that it becomes bivariate. This allows you to
change the lag. (defaults to \code{1})}
}
\value{
an \code{array} of dimension \code{c(num_datasets, 1000, 2)} which can be input
into \code{nnadic()} for prediction.
}
\description{
Input your bivariate data and get out the 500 most extreme (by l-infinity
norm) points.
The function defaults to providing 100 datasets. These 100 datasets can be
copies if your data has exactly 10,000 points or if \code{subsample = FALSE} and
your data have more than 10,000 points without l-infinity norm ties at the
0.95 quantile. If there are ties in the previous cases then the 100
returned data sets will differ on a small number of points. These points
are the tied points which are sampled for inclusion in each dataset.
If your data have less than 10,000 points the top five percent are
re-sampled up to 500 points 100 different times.
Note that all datasets will show 1000 points as the 500 largest points are
repeated, with reflection, to enforce symmetry in all datsets. This makes
all \code{nnadic()} predictions invariant to components of the random vector.
}
\examples{
my_favorite_data <- matrix(rnorm(20000), nrow = 10000)
get_nnadic_input(my_favorite_data)
}
