% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roc.R
\name{get_roc}
\alias{get_roc}
\title{Compute the ROC (receiver operating characteristic) curve for 200 evenly
spaced thresholds}
\usage{
get_roc(predicted_probabilities, truth, make_plot = FALSE, add_auc = FALSE)
}
\arguments{
\item{predicted_probabilities}{the "probabilities" returned from \code{nnadic()}}

\item{truth}{the true classifications (a vector of zeros and ones)}

\item{make_plot}{set to \code{TRUE} if you want an ROC plot to be auto-generated.
Default is \code{FALSE}}

\item{add_auc}{set to \code{TRUE} if you want the AUC (area under the curve)
added to the ROC plot. Default is \code{FALSE}. Does nothing if set to \code{TRUE}
while \code{make_plot = FALSE}.}
}
\value{
a matrix of false and true positive rates (FPR, TPR)
}
\description{
Compute the ROC (receiver operating characteristic) curve for 200 evenly
spaced thresholds
}
\examples{
library(nnadicTestData)
test_data_sym <- make_symmetric(test_data_four)
predicted_probs <- nnadic(test_data_sym)$probs
roc <- get_roc(predicted_probabilities = predicted_probs,
               truth = test_response_four,
               make_plot = TRUE,
               add_auc = TRUE)
}
