% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_marginal.R
\name{make_symmetric}
\alias{make_symmetric}
\title{Make the data symmetric by copying points with swapped coordinates}
\usage{
make_symmetric(data)
}
\arguments{
\item{data}{the matrix or array including dataset(s) of the largest 500 points}
}
\value{
an array with 1000 points for each dataset.
}
\description{
This function takes the 500 largest points $(X,Y)$ and appends the same
points after reflection across the identity line $(Y, X)$. This ensures all
input datasets are symmetric and the \code{nnadic()} classification probabilities
are invariant to the order of the components in the random vector.
}
\examples{
largest_500 <- matrix(rnorm(1000), nrow = 500)
symmetric_nnadic_input <- make_symmetric(largest_500)
}
