% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnadic.R
\name{nnadic}
\alias{nnadic}
\title{Neural Network for Asymptotic Dependence/Independence Classification}
\usage{
nnadic(data)
}
\arguments{
\item{data}{a matrix of dimension \eqn{d}x\eqn{2}.
Currently only supports \eqn{d=500}}
}
\value{
a \code{list} with two items:
\enumerate{
\item probs - a vector of probabilities of being AI
\item preds - a vector of predictions (0 is AD, 1 is AI)
}
}
\description{
Neural Network for Asymptotic Dependence/Independence Classification
}
\examples{
nnadic(matrix(rnorm(1000), ncol = 2))
}
