% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_marginal.R
\name{resample_to_500}
\alias{resample_to_500}
\title{Get indices for the 500 points to be used as \code{nnadic} input}
\usage{
resample_to_500(
  indices,
  tie_indices. = NULL,
  subsample. = TRUE,
  num_datasets. = 100,
  include_all. = TRUE,
  verbose. = TRUE
)
}
\arguments{
\item{indices}{a vector with of indices of large points}

\item{tie_indices.}{vector of indices for tied values at the
large-point-cutoff}

\item{subsample.}{Set to \code{FALSE} if the dataset has at least than 10,000
points and you want to take the top 500 points rather than subsample all
points above the 0.95 quantile. (defaults to \code{TRUE})}

\item{num_datasets.}{The number of output datasets (defaults to \code{100})}

\item{include_all.}{If there are fewer than 10,000 points in your data should
all of the top five percent be included in every dataset as many times as
possible (\code{TRUE}) or should all 500 points be from resampling (\code{FALSE})?
(defaults to \code{TRUE})}

\item{comp_lag}{If your data is a univariate time series the function will
automatically lag it so that it becomes bivariate. This allows you to
change the lag. (defaults to \code{1})}

\item{verbose}{Switch to \code{FALSE} to skip comments. (defaults to \code{TRUE})}
}
\value{
a \code{matrix} of dimension \code{c(500, num_datasets)} containing indices
of large points after sub/re-sampling
}
\description{
Input your large indices and get out the appropriate sub/re-sampled indices.
}
\examples{
sample_inds <- sample(1:9000, size = 450)
resample_to_500(sample_inds, num_datasets. = 1)
}
