% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_marginal.R
\name{transform_to_exponential}
\alias{transform_to_exponential}
\title{Transform marginal distribution to unit exponential}
\usage{
transform_to_exponential(data)
}
\arguments{
\item{data}{a vector of observations}
}
\value{
a vector with unit exponential marginal distribution
}
\description{
Transform marginal distribution to unit exponential
}
\examples{
exp_margins[,1] <- transform_to_exponential(my_favorite_data[,1])
}
