% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roc.R
\name{get_tpr_fpr}
\alias{get_tpr_fpr}
\title{Computes the true positive rate and false positive rate from a confusion
matrix}
\usage{
get_tpr_fpr(conf_mat)
}
\arguments{
\item{conf_mat}{confusion matrix, often the output from
\code{get_confusion_matrix()}}
}
\value{
a vector with two values, the false positive rate (fpr) and the true
positive rate (tpr) in that order.
}
\description{
Computes the true positive rate and false positive rate from a confusion
matrix
}
\examples{
confusion_matrix <- get_confusion_matrix(pred_prob = predicted_prob,
                                         truth = test_response)
get_tpr_fpr(conf_mat = confusion_matrix)
}
