% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnadic.R
\name{nnadic}
\alias{nnadic}
\title{Neural Network for Asymptotic Dependence/Independence Classification}
\usage{
nnadic(data, one_test = TRUE, make_hist = TRUE)
}
\arguments{
\item{data}{an array of dimension \eqn{n}x\eqn{d}x\eqn{2}.
Currently only supports \eqn{d=1000}. Use \code{get_nnadic_input()} to prepare
your data.}

\item{one_test}{change to \code{FALSE} if there are multiple datasets in your test
This can be used to test nnadic on simulated datasets where the truth is
known or on multiple unknown datasets at the same time.}

\item{make_hist}{change to \code{FALSE} if you do not want a histogram of the
probabilities to be returned.}
}
\value{
a \code{list} with two or three items:
\enumerate{
\item probs - a vector of probabilities of being AI
\item preds - a vector of predictions (0 is AD, 1 is AI)
\item mean  - the mean of the predictions, included when one_test is TRUE
}
}
\description{
Neural Network for Asymptotic Dependence/Independence Classification
}
\examples{
nnadic(get_nnadic_input(matrix(rnorm(20000), ncol = 2)))
}
