% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_marginal.R
\name{transform_to_exponential}
\alias{transform_to_exponential}
\title{Transform marginal distribution to unit exponential}
\usage{
transform_to_exponential(data, verbose = TRUE)
}
\arguments{
\item{data}{a vector of observations}

\item{verbose}{Switch to \code{FALSE} to skip comments. (defaults to \code{TRUE})}
}
\value{
a vector with unit exponential marginal distribution
}
\description{
Transform marginal distribution to unit exponential
}
\examples{
my_favorite_data <- matrix(rnorm(20000), nrow = 10000)
exp_margins <- my_favorite_data
exp_margins[,1] <- transform_to_exponential(my_favorite_data[,1])
}
